Bug fixes, and breaking out the classification pass as a saved step
